pyproject.toml
[tool.poetry]
name = "social_network"
version = "0.1.0"
description = ""
authors = ["Aleksandr <amvorobei@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
django-debug-toolbar = "^3.2.4"
djangorestframework = "^3.13.1"
django-rest-swagger = "^2.2.0"
djangorestframework-simplejwt = "^5.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
